+++ python/lsst/ts/mtdomegui/tab/tab_elevation.py
@@ -151,7 +151,7 @@ class TabElevation(TabTemplate):
             "drive_torque": TabFigure(
                 "Actual Drive Torque",
                 self.model,
-                "J",
+                "N*m",
                 [str(idx) for idx in range(LWSCS_NUM_MOTORS)],
             ),
             "drive_current": TabFigure(
@@ -398,10 +398,10 @@ class TabElevation(TabTemplate):
 
         for idx in range(LWSCS_NUM_MOTORS):
             self._status["drive_torque_commanded"][idx].setText(
-                f"{telemetry['driveTorqueCommanded'][idx]:.2f} J"
+                f"{telemetry['driveTorqueCommanded'][idx]:.2f} N*m"
             )
             self._status["drive_torque_actual"][idx].setText(
-                f"{telemetry['driveTorqueActual'][idx]:.2f} J"
+                f"{telemetry['driveTorqueActual'][idx]:.2f} N*m"
             )
             self._status["drive_current_actual"][idx].setText(
                 f"{telemetry['driveCurrentActual'][idx]:.2f} A"