+++ tests/tab/test_tab_calibration.py
@@ -69,8 +69,8 @@ async def test_set_signal_telemetry(widget: TabCalibration) -> None:
     assert widget._status["position_commanded"].text() == "1.00"
     assert widget._status["position_actual"].text() == "1.00"
 
-    assert widget._status["drive_torque_commanded"].text() == "1.00 J"
-    assert widget._status["drive_torque_actual"].text() == "1.00 J"
+    assert widget._status["drive_torque_commanded"].text() == "1.00 N*m"
+    assert widget._status["drive_torque_actual"].text() == "1.00 N*m"
     assert widget._status["drive_current_actual"].text() == "1.00 A"
 
     assert widget._status["drive_temperature"].text() == "1.00 deg C"