+++ doc/version_history.rst
@@ -14,6 +14,7 @@ Version History
 
 * Allow to query the shutter status.
 * Fix the unit of position of the shutter.
+* Change the torque unit to be N*m instead of J.
 
 .. _lsst.ts.mtdomegui-0.4.2: